requirement.txt
# Core dependencies
gitpython>=3.1.30
python-dotenv>=1.0.0
requests>=2.28.2
tqdm>=4.65.0

# Code analysis
astroid>=2.15.0
libcst>=1.0.0
tree-sitter>=0.20.1

# NLP & ML
transformers>=4.27.0
sentence-transformers>=2.2.2
numpy>=1.24.2
scikit-learn>=1.2.2

# Vector database
pinecone-client>=2.2.1
faiss-cpu>=1.7.4
pymilvus>=2.2.8

# Evaluation metrics
rouge-score>=0.1.2
bert-score>=0.3.13
nltk>=3.8.1
py-rouge>=1.1
sacrebleu>=2.3.1

# Utilities
pydantic>=2.0.0
rich>=13.3.5
typer>=0.9.0
click>=8.1.3

POT=0.9.5

#Search
langchain==0.1.14
langchain-community==0.3.23
openai==1.76.2
tiktoken==0.9.0
